♻️ (Refactor) - Large-scale refactoring after beta publishing
