+++ src/lib/interfaces/IDocumentationViewActionParams.ts
@@ -1,0 +1,12 @@
+import * as vscode from "vscode";
+import { IDocumentation } from "./IDocumentation";
+
+interface IDocumentationViewActionParams {
+  id: string;
+  documentations: IDocumentation[];
+  extensionUri: vscode.Uri;
+  panels: { [id: string]: vscode.WebviewPanel };
+  webview: vscode.Webview;
+}
+
+export default IDocumentationViewActionParams;
